pyproject.toml
[project]
name = "prismberry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.5",
    "fastsql>=2.0.3",
    "pillow>=11.0.0",
    "python-fasthtml>=0.10.1",
    "ruff>=0.8.2",
    "sqlmodel>=0.0.22",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipython>=8.30.0",
    "jupyter>=1.1.1",
]
